pyproject.toml
[tool.poetry]
name = "BourseVisualisation"
version = "0.1.0"
description = ""
authors = ["Leandre Tuesta <59528877+TLeandre@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.10"
notebook = "^6.5.2"
yfinance = "^0.2.12"
matplotlib = "^3.7.0"
streamlit = "^1.18.1"
plotly = "^5.13.0"
pandas-datareader = "^0.10.0"
streamlit-option-menu = "^0.3.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
